% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generatePageURLs.R
\name{generatePageURLs}
\alias{generatePageURLs}
\title{Generate URLs of the database overview pages}
\usage{
generatePageURLs(pages)
}
\arguments{
\item{pages}{The number of page URLs to return}
}
\value{
(character vector) The desired number of URLs
}
\description{
This function generates a speciefied number of database pages for the
project databse of the IFC. First page here
http://ifcextapps.ifc.org/ifcext/spiwebsite1.nsf/frmShowView?openform&view=CRUDate&start=1&count=100&page=1
The returned pages start at page one and end at page 'pages'.
}

