% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getProjectInfo.R
\name{getProjectInfo}
\alias{getProjectInfo}
\title{Get the project information from a project's HTML}
\usage{
getProjectInfo(projectHTML, tabNames)
}
\arguments{
\item{projectHTML}{Usually the result of rvest::read_html()}

\item{tabNames}{(character) Usually the result of getTabNames()}
}
\value{
A list with the pieces of information about the project as
its elements
}
\description{
This is the main function of the IFCscraper package. It returns project
information given a project's HTML (from read_html) and the tabNames of the
project's page.
}

