% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getProjLinksFromPage.R
\name{getProjLinksFromPage}
\alias{getProjLinksFromPage}
\title{Get all URLs of projects from a database page}
\usage{
getProjLinksFromPage(URL)
}
\arguments{
\item{URL}{(character) The URL of an overview page of the IFC's database}
}
\value{
(character vector) All found project links
}
\description{
This function takes a URL of an 'overview page' of the IFC's database
(like http://ifcextapps.ifc.org/ifcext/spiwebsite1.nsf/frmShowView?openform&view=CRUDate&start=1&count=100&page=1)
as input and returns the URLs of all projects on that site.
}

