% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTabNames.R
\name{getTabNames}
\alias{getTabNames}
\title{Get the names of the tabs at the bottom of a project's page}
\usage{
getTabNames(projectHTML)
}
\arguments{
\item{projectHTML}{Usually the output of readr::read_html() applied to the
URL of a single project}
}
\value{
(character vector) The names of the tabs at the bottom of the page
}
\description{
This function returns the names of the tabs (if any) that may be at the
bottom of a project's page.
}

