\name{IFCprojects}
\docType{data}
\alias{IFCprojects}
\title{The returned project information without merging by project ID}
\description{
  This data set contains the result of scraping all 68 pages of the IFC's database
  (October 2015). The rows can be combined by merging them using the project number.
  The merged data set is in IFCprojectsMerged.
}
\usage{IFCprojects}
\format{A data frame with 6777 obs. of  12 variables}
\keyword{datasets}