\name{IFCprojectsMerged}
\docType{data}
\alias{IFCprojectsMerged}
\title{The returned project information after merging by project ID}
\description{
  This data set contains the result of scraping all 68 pages of the IFC's database
  (October 2015). The rows have already been combined by merging them using the project number.
  There were conflicts in projects 26504, 27091, 27373, 27373, 28741, 37243, and 8648 which
  lead to a loss of information. The conflicts have to be checked manually. 
}
\usage{IFCprojectsMerged}
\format{A data frame with 4823 obs. of  12 variables}
\keyword{datasets}